[BOT] refactor(rename): InputStream_Sub1 -> MuLawInputStream
